[Jinhan] placeListView query and search

+++ core/src/main/kotlin/com/tamerofficial/common/Location.kt
@@ -1,5 +1,7 @@
 package com.tamerofficial.common
 
+import java.math.BigDecimal
+
 data class Location (
-    val lat : Long, val lon : Long
+    val lat : BigDecimal, val lon : BigDecimal
 )
